refactor: remove metametrics circular dependency
