+++ shared/modules/metametrics.test.ts
@@ -4,10 +4,8 @@ import {
   TransactionType,
 } from '@metamask/transaction-controller';
 import { createTestProviderTools } from '../../test/stub/provider';
-// TODO: Remove restricted import
-// eslint-disable-next-line import/no-restricted-paths
-import { TransactionMetricsRequest } from '../../app/scripts/lib/transaction/metrics';
 import { CHAIN_IDS } from '../constants/network';
+import { TransactionMetricsRequest } from '../types/metametrics';
 import { getSmartTransactionMetricsProperties } from './metametrics';
 
 const txHash =
